refactor(svelte2tsx): 要素ノードのハンドラを nodes/ へ切り出す
element (通常要素と <title>) / dynamic_element (<svelte:element>) /
special_element (svelte: メタ要素) を純粋移動。出力・ロジックの変更なし。
