+++ crates/rsvelte_core/src/svelte2tsx/template/mod.rs
@@ -16,8 +16,7 @@ mod walk;
 
 use crate::ast::template::{
     Attribute, AttributeValue, AttributeValuePart, Component, Fragment, LetDirective,
-    RegularElement, SlotElement, SvelteComponentElement, SvelteDynamicElement, SvelteElement,
-    TemplateNode, TitleElement,
+    RegularElement, SlotElement, SvelteComponentElement, SvelteElement, TemplateNode,
 };
 use std::fmt::Write as _;
 
@@ -27,23 +26,14 @@ use super::magic_string::MagicString;
 use super::svelte2tsx::{Svelte2TsxOptions, SvelteVersion, slice_src};
 use attributes::action::format_use_directive;
 use attributes::attribute::format_attribute_node;
-use attributes::binding::{
-    any_bind_needs_element_var, build_bind_directive_suffix, element_var_base_name,
-    format_bind_directive, sanitize_tag_for_var,
-};
+use attributes::binding::format_bind_directive;
 use attributes::class_style::build_class_style_directive_suffix_segments;
-use attributes::directive_suffix::{
-    build_component_directive_suffix, build_directive_prefix_suffix,
-    build_element_directive_suffix_segments,
-};
+use attributes::directive_suffix::build_component_directive_suffix;
 use attributes::event_handler::{build_on_calls, format_on_directive, get_on_directives};
 use attributes::let_::{build_let_destructure_string, get_let_directives};
 use attributes::spread::format_spread_attribute;
 use attributes::transition::format_transition_directive;
-use attributes::{
-    build_attribute_segments, build_attributes_string, build_component_props_segments,
-    build_component_props_string,
-};
+use attributes::{build_component_props_segments, build_component_props_string};
 use ctx::{Counter, TemplateNodeExt};
 use segs::{
     Seg, bake_out_of_order_src, emit_segmented_overwrite, segs_is_empty, segs_to_string,
@@ -54,14 +44,11 @@ use utils::expr::{
     get_expression_range, get_expression_text, get_set_binding_ranges,
 };
 use utils::names::{reversed_component_instance_name, reversed_component_name};
-use utils::source::{
-    closing_tag_name_matches, count_tag_to_attr_spaces, find_closing_tag_start,
-    find_opening_tag_end,
-};
+use utils::source::{count_tag_to_attr_spaces, find_closing_tag_start, find_opening_tag_end};
 
 pub(crate) use ctx::{clear_element_opener_comments, set_element_opener_comments};
 use nodes::attach_tag::format_attach_tag_segments;
-use nodes::snippet_block::{handle_snippet_block_as_component_prop, hoist_snippet_blocks};
+use nodes::snippet_block::handle_snippet_block_as_component_prop;
 use walk::{process_fragment_inplace, process_node_inplace};
 
 // =============================================================================
@@ -157,255 +144,6 @@ pub fn collect_template_info(fragment: &Fragment, source: &str) -> TemplateInfo 
 // =============================================================================
 // Element Nodes
 // =============================================================================
-
-/// Handle a regular HTML element.
-///
-/// Generates `{ svelteHTML.createElement("tagName", { ...attributes }); children }`.
-///
-/// The opening tag `<h1 class="foo">` is overwritten with
-/// `{ svelteHTML.createElement("h1", {"class":\`foo\`,});`
-/// and the closing tag `</h1>` is overwritten with ` }`.
-fn handle_regular_element(
-    el: &RegularElement,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    if el.start >= el.end {
-        return;
-    }
-
-    // A nested `<style>` element is removed entirely from the output,
-    // mirroring official svelte2tsx's `handleStyleTag` (the `case 'Style'`
-    // arm), which does `str.remove(node.start, node.end)` for every verbatim
-    // style node at any nesting depth. (A top-level `<style>` becomes
-    // `root.css` and never reaches this fragment walk, so any `style`
-    // RegularElement here is necessarily nested.) Note: nested `<script>`
-    // elements are NOT removed — official emits `createElement("script", {})`
-    // for them (only the JS content is blanked, which `handle_text` already
-    // does), so they fall through to the normal element path.
-    if el.name == "style" {
-        str.remove(el.start, el.end);
-        return;
-    }
-
-    // Official svelte2tsx switches the opener on the *tag name*, not the AST node
-    // type: any element named `slot` emits `__sveltets_createSlot(...)`. The parser
-    // only produces a `SlotElement` for `<slot>` outside a `<template
-    // shadowrootmode>`; inside one it is a `RegularElement` (mirroring upstream's
-    // `parent_is_shadowroot_template` check), yet svelte2tsx still lowers it to a
-    // slot. Route those through the same slot handler.
-    if el.name == "slot" {
-        let slot = SlotElement {
-            start: el.start,
-            end: el.end,
-            name: el.name.clone(),
-            name_loc: el.name_loc,
-            attributes: el.attributes.clone(),
-            fragment: el.fragment.clone(),
-        };
-        handle_slot_element(&slot, source, options, str, counter, depth);
-        return;
-    }
-
-    // Named-slot routing: when processing a component's children (possibly deep
-    // inside `{#each}`/`{#if}`/etc. control-flow blocks), an element targeting a
-    // named slot is lowered to the `$$slot_def[...]` form referencing the
-    // enclosing component instance. Take the context first so this element's OWN
-    // children do NOT inherit it (a nested element owns its own slot scope);
-    // restore it afterwards for the following siblings.
-    let saved_slot = counter.slot_inst.take();
-    if let Some(ref inst) = saved_slot
-        && get_slot_attr_value(&el.attributes, source).is_some()
-    {
-        handle_named_slot_element(el, inst, source, options, str, counter, depth);
-        counter.slot_inst = saved_slot;
-        return;
-    }
-
-    // Find the end of the opening tag (after the `>`)
-    let opening_tag_end = find_opening_tag_end(source, el.start, el.end);
-
-    // Build attribute segments. Source-bearing expressions become
-    // `Seg::Src` so the resulting overwrite leaves them as unedited
-    // MagicString chunks — which `generate_mappings` then maps
-    // per-character back to the original `.svelte` columns. Element-
-    // opener attribute expressions previously baked into a single
-    // edited chunk and collapsed to a single source-map segment.
-    // `saved_slot` (taken from `counter.slot_inst` above) is Some when this
-    // element is a slot-context child of a component — then `let:` is a slot-let,
-    // not a regular attribute.
-    // The opener content (where attributes + comments live) starts right after
-    // `<tagname`, so leading comments before the first attribute are recovered.
-    let opener_content_start = el.start + 1 + el.name.len() as u32;
-    let mut attr_segs = build_attribute_segments(
-        &el.attributes,
-        source,
-        &el.name,
-        saved_slot.is_some(),
-        Some(opener_content_start),
-    );
-
-    // Official always emits exactly ONE inherent space after the `{` of the
-    // attribute object, regardless of the source whitespace between the tag name
-    // and the first attribute (verified: `<button onclick>`, `<button  onclick>`,
-    // `<button\n\tonclick>` all → `{ "onclick":… }`). oxfmt normalises this away
-    // for valid output, but a raw top-level-await component keeps it exact.
-    let attrs_empty_before_pad = segs_is_empty(&attr_segs);
-    if !el.attributes.is_empty() && !attrs_empty_before_pad {
-        segs_trim_start(&mut attr_segs);
-        let mut padded: Vec<Seg> = Vec::with_capacity(attr_segs.len() + 1);
-        padded.push(Seg::Lit(" ".to_string()));
-        padded.extend(attr_segs);
-        attr_segs = padded;
-    }
-
-    // V4-style action / transition / animate directive emission. Action
-    // becomes `const $$action_N = __sveltets_2_ensureAction(…);` BEFORE
-    // the createElement; transition / animate become
-    // `__sveltets_2_ensureTransition(…);` appended AFTER it. The
-    // createElement's second argument also needs to wrap any actions
-    // with `__sveltets_2_union(...)`. Mirrors
-    // `htmlxtojsx_v2/nodes/{Action,Transition,Animation}.ts`.
-    // Only the action PREFIX (`const $$action_N = …`) and the action count are
-    // taken here; the transition/animate suffix is emitted in source order by
-    // `build_element_directive_suffix_segments` below.
-    let (directive_prefix, _directive_suffix, action_count) =
-        build_directive_prefix_suffix(&el.attributes, source, &el.name);
-    let actions_arg = if action_count > 0 {
-        let mut args = String::from(", __sveltets_2_union(");
-        for i in 0..action_count {
-            if i > 0 {
-                args.push(',');
-            }
-            let _ = write!(args, "$$action_{}", i);
-        }
-        args.push(')');
-        args
-    } else {
-        String::new()
-    };
-
-    // `bind:` directives generate a suffix appended right after the
-    // createElement call. Mirrors `htmlxtojsx_v2/nodes/Binding.ts::handleBinding`.
-    // For `bind:this` and one-way bindings on the element (`offsetHeight`,
-    // …) we also need a `const $$_xxx = …` declaration so the assignment
-    // can reference the element value.
-    let needs_element_var = any_bind_needs_element_var(&el.attributes, source);
-    let element_var = if needs_element_var {
-        // The `$$_<tag><N>` index is the element's nesting DEPTH (matching
-        // upstream Element.ts `computeDepth()`), not a per-tag counter — same
-        // rule as component instance names.
-        let sanitized = sanitize_tag_for_var(&el.name);
-        Some(format!("$$_{}{}", sanitized, depth))
-    } else {
-        None
-    };
-    // All post-`createElement` directive statements — `class:` / `style:`
-    // (segmented), `transition:` / `in:` / `out:` / `animate:`, and `bind:` —
-    // are built in a SINGLE source-order pass so they interleave exactly like
-    // official's `appendToStartEnd` walk (e.g. a `style:` after a `bind:this`
-    // stays after it instead of grouping with earlier `class:` directives).
-    let suffix_segs = build_element_directive_suffix_segments(
-        &el.attributes,
-        source,
-        element_var.as_deref(),
-        &el.name,
-        options.is_ts_file,
-        &el.name,
-    );
-
-    // When all surviving props are empty but a `bind:` / `class:` / `style:`
-    // directive was stripped, JS reference still leaves whitespace inside
-    // `{ }`. Add a single space so `createElement("div", { })` matches.
-    if segs_is_empty(&attr_segs) && !segs_is_empty(&suffix_segs) {
-        attr_segs.push(Seg::Lit(" ".into()));
-    }
-
-    // Build the opener as a `Vec<Seg>` (header lit + attr segs + trailer
-    // lit) and apply via `emit_segmented_overwrite`. Action declarations
-    // (if any) are emitted *before* the inner `{ … createElement(…); … }`
-    // block so they're in scope for `__sveltets_2_union(...)`. The inner
-    // `{` opens a separate block scope.
-    let element_var_decl = if let Some(ref var) = element_var {
-        format!("const {} = ", var)
-    } else {
-        String::new()
-    };
-    let header_lit = if !directive_prefix.is_empty() {
-        format!(
-            " {{{}{{ {}svelteHTML.createElement(\"{}\"{}, {{",
-            directive_prefix, element_var_decl, el.name, actions_arg,
-        )
-    } else {
-        format!(
-            " {{ {}svelteHTML.createElement(\"{}\"{}, {{",
-            element_var_decl, el.name, actions_arg,
-        )
-    };
-    // The trailer closes the props object + createElement call (`}});`), then
-    // appends the `class:` / `style:` directive statements (segmented, so their
-    // expression chunks keep their source mapping), then the transition/animate
-    // (`directive_suffix`) and `bind:` (`bind_suffix`) suffixes.
-    let mut opener_segs: Vec<Seg> = Vec::with_capacity(attr_segs.len() + suffix_segs.len() + 3);
-    opener_segs.push(Seg::Lit(header_lit));
-    opener_segs.extend(attr_segs);
-    // Close the props object + createElement call: `});` (one `}` for the
-    // props brace, then `)` + `;`). The outer block `{` is closed after the
-    // children by the closing-tag overwrite.
-    opener_segs.push(Seg::Lit("});".to_string()));
-    // The post-`createElement` suffix statements are already assembled in
-    // source-attribute order by `build_element_directive_suffix_segments`.
-    opener_segs.extend(suffix_segs);
-    let opener_segs = bake_out_of_order_src(opener_segs, source);
-    emit_segmented_overwrite(str, el.start, opening_tag_end, &opener_segs);
-
-    // Process children at depth+1: this element is now an ancestor.
-    // Mirrors official computeDepth which counts all ancestor element/component nodes.
-    // Hoist snippet blocks to the top of the element's children first, mirroring
-    // hoistSnippetBlock in the JS reference (pendingSnippetHoistCheck walk).
-    hoist_snippet_blocks(&el.fragment, source, str);
-    process_fragment_inplace(&el.fragment, source, options, str, counter, depth + 1);
-
-    // Find and overwrite the closing tag.
-    // HTML void elements (`<input>`, `<br>`, …) and source-level self-closing
-    // tags (`<x />`) have no `</tag>` in the source, so we must NOT call
-    // `find_closing_tag_start` on them — it scans backwards for `</` and would
-    // wrongly match a preceding sibling's closing tag, blanking it (and the
-    // void element itself) on overwrite. Mirrors the JS reference's
-    // `prependLeft(node.end, '}')` for void/self-closing tags.
-    //
-    // When `directive_prefix` opened an extra outer block for the action
-    // declarations, emit a matching extra `}` to close it.
-    let extra_close = if directive_prefix.is_empty() { "" } else { "}" };
-    let is_self_closing_source = slice_src(source, el.start as usize, el.end as usize)
-        .trim_end()
-        .ends_with("/>");
-    let is_void = crate::compiler::utils::is_void_element(&el.name);
-    if is_void || is_self_closing_source {
-        str.append_left(el.end, &format!("}}{}", extra_close));
-    } else {
-        let closing_tag_start = find_closing_tag_start(source, el.end);
-        // An auto-closed element (`<p><p>`, `<li><li>`, …) has NO `</name>` at
-        // `el.end`; `find_closing_tag_start` then wrongly matches the last
-        // child's `</…>`. Only overwrite when the found tag actually closes
-        // THIS element; otherwise append `}` at `el.end` like a void element
-        // (matching official's `prependLeft(node.end, '}')` for such cases).
-        if closing_tag_start < el.end
-            && closing_tag_name_matches(source, closing_tag_start, &el.name)
-        {
-            // Non-self-closing: preserve space before closing brace
-            str.overwrite(closing_tag_start, el.end, &format!(" }}{}", extra_close));
-        } else {
-            str.append_left(el.end, &format!("}}{}", extra_close));
-        }
-    }
-    // Restore the slot context for following siblings (this element's own
-    // children were processed with it cleared, via the `take()` above).
-    counter.slot_inst = saved_slot;
-}
 
 /// Handle a Svelte component: `<Component ...>`.
 ///
@@ -1620,270 +1358,6 @@ fn handle_svelte_component(
     counter.slot_inst = saved_outer_slot;
 }
 
-/// Handle `<svelte:element this={tag}>`.
-fn handle_svelte_dynamic_element(
-    el: &SvelteDynamicElement,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    if el.start >= el.end {
-        return;
-    }
-
-    // Named-slot routing: `<svelte:element … slot="x">` inside a component's
-    // children targets the parent component's named slot. Wrap the whole
-    // `createElement(...)` in a `$$slot_def["x"]` block and drop the `slot`
-    // attribute. Take the context so the element's own children don't inherit
-    // it; restore it for following siblings.
-    let saved_slot = counter.slot_inst.take();
-    let named_slot: Option<(String, String)> = saved_slot.as_ref().and_then(|inst| {
-        get_slot_attr_value(&el.attributes, source).map(|name| (inst.clone(), name))
-    });
-    if let Some((ref inst, ref target_slot)) = named_slot {
-        let lets = get_let_directives(&el.attributes);
-        let let_destructure = build_let_destructure_string(&lets, source);
-        let block_open = format!(
-            "{{const {{/*\u{03A9}ignore_start\u{03A9}*/$$_$$/*\u{03A9}ignore_end\u{03A9}*/,{}}} = {}.$$slot_def[\"{}\"];$$_$$;",
-            let_destructure, inst, target_slot
-        );
-        str.prepend_left(el.start, &block_open);
-    }
-
-    let raw_tag_text = get_expression_text(&el.tag, source);
-    // If the `this` attribute value is a plain string literal (this="tag"),
-    // the parser stores just the text without quotes. We need to wrap it
-    // in quotes to produce valid JavaScript: createElement("tag", ...).
-    let tag_text = if let Some((start, _end)) = get_expression_range(&el.tag) {
-        let before = if start > 0 {
-            source.as_bytes()[(start - 1) as usize]
-        } else {
-            b'{'
-        };
-        if before == b'"' || before == b'\'' {
-            // String literal: wrap in quotes
-            format!("\"{}\"", raw_tag_text)
-        } else {
-            raw_tag_text.to_string()
-        }
-    } else {
-        raw_tag_text.to_string()
-    };
-    let opening_tag_end = find_opening_tag_end(source, el.start, el.end);
-    // In a named-slot context the `slot` attribute is consumed by the wrapper
-    // block, so build the attributes without it.
-    let attrs_str = if named_slot.is_some() {
-        build_named_slot_element_attrs(&el.attributes, source)
-    } else {
-        build_attributes_string(&el.attributes, source, saved_slot.is_some())
-    };
-
-    // `use:` / `transition:` / `animate:` directives, same V4 emission as on a
-    // regular element. The action's `mapElementTag` uses the literal element
-    // name (`svelte:element`); the `createElement` first arg stays the dynamic
-    // tag expression.
-    let (directive_prefix, directive_suffix, action_count) =
-        build_directive_prefix_suffix(&el.attributes, source, &el.name);
-    let actions_arg = if action_count > 0 {
-        let mut args = String::from(", __sveltets_2_union(");
-        for i in 0..action_count {
-            if i > 0 {
-                args.push(',');
-            }
-            let _ = write!(args, "$$action_{}", i);
-        }
-        args.push(')');
-        args
-    } else {
-        String::new()
-    };
-    // Only the action `directive_prefix` (the `const $$action_N = …;`
-    // declarations) needs an extra inner block scope; a transition/animate-only
-    // suffix is just appended after the createElement, no extra braces.
-    let needs_inner_block = !directive_prefix.is_empty();
-
-    // Check if this is a self-closing element (no separate closing tag).
-    // Also covers HTML void elements like `<input>`, `<br>`, `<img>` which have
-    // no closing tag in the source — `is_void_element` keeps the opener and
-    // closing brace on a single line, mirroring the JS reference's behaviour
-    // for void tags.
-    let is_self_closing = el.fragment.nodes.is_empty()
-        && (slice_src(source, el.start as usize, el.end as usize)
-            .trim_end()
-            .ends_with("/>")
-            || crate::compiler::utils::is_void_element(&el.name));
-
-    let attrs_self = if attrs_str.is_empty() {
-        "  "
-    } else {
-        &attrs_str
-    };
-    let attrs_open = if attrs_str.is_empty() {
-        " "
-    } else {
-        &attrs_str
-    };
-    // With directives an extra inner block scope wraps the createElement so the
-    // action declarations (in `directive_prefix`) are in scope: ` {<prefix>{ … }}`.
-    let inner_open = if needs_inner_block { "{" } else { "" };
-    let inner_close = if needs_inner_block { "}" } else { "" };
-    // `bind:this` / one-way bindings on `<svelte:element>` need the
-    // `const $$_svelteelement<depth> = createElement(...)` form so the binding
-    // assignment can reference it. Mirrors regular-element / Element.ts lowering.
-    let needs_element_var = any_bind_needs_element_var(&el.attributes, source);
-    let element_var = if needs_element_var {
-        Some(format!("$$_{}{}", element_var_base_name(&el.name), depth))
-    } else {
-        None
-    };
-    let bind_suffix = build_bind_directive_suffix(
-        &el.attributes,
-        source,
-        element_var.as_deref(),
-        &el.name,
-        options.is_ts_file,
-    );
-    let element_var_decl = element_var
-        .as_ref()
-        .map(|v| format!("const {} = ", v))
-        .unwrap_or_default();
-    // `class:`/`style:` directives lower to statements after the createElement
-    // (`class:active={x}` → ` x;`), same as a regular element.
-    let class_style_suffix = segs_to_string(
-        &build_class_style_directive_suffix_segments(&el.attributes, source),
-        source,
-    );
-    // ` <var=>svelteHTML.createElement(tag<actions_arg>, {attrs});<suffix>` — no
-    // leading `{`; the block brace comes from the outer ` {` (and `inner_open`
-    // when directives add an extra scope).
-    // The post-`createElement` suffix statements — `class:`/`style:`, transition/animate
-    // (`directive_suffix`), and `bind:` (`bind_suffix`) — are emitted in SOURCE-ATTRIBUTE
-    // ORDER, mirroring the regular-element handler's sort logic.
-    let first_bind_pos_se = el
-        .attributes
-        .iter()
-        .filter_map(|a| match a {
-            Attribute::BindDirective(b) => Some(b.start),
-            _ => None,
-        })
-        .min();
-    let first_directive_pos_se = el
-        .attributes
-        .iter()
-        .filter_map(|a| match a {
-            Attribute::TransitionDirective(t) => Some(t.start),
-            Attribute::AnimateDirective(an) => Some(an.start),
-            _ => None,
-        })
-        .min();
-    let first_class_style_pos_se = el
-        .attributes
-        .iter()
-        .filter_map(|a| match a {
-            Attribute::ClassDirective(c) => Some(c.start),
-            Attribute::StyleDirective(s) => Some(s.start),
-            _ => None,
-        })
-        .min();
-    let sorted_suffix = {
-        let mut pieces: Vec<(u32, &str)> = Vec::new();
-        if !directive_suffix.is_empty() {
-            pieces.push((
-                first_directive_pos_se.unwrap_or(u32::MAX),
-                &directive_suffix,
-            ));
-        }
-        if !class_style_suffix.is_empty() {
-            pieces.push((
-                first_class_style_pos_se.unwrap_or(u32::MAX),
-                &class_style_suffix,
-            ));
-        }
-        if !bind_suffix.is_empty() {
-            pieces.push((first_bind_pos_se.unwrap_or(u32::MAX), &bind_suffix));
-        }
-        pieces.sort_by_key(|(pos, _)| *pos);
-        pieces.into_iter().map(|(_, s)| s).collect::<String>()
-    };
-    let create = |attrs: &str| {
-        format!(
-            " {}svelteHTML.createElement({}{}, {{{}}});{}",
-            element_var_decl, tag_text, actions_arg, attrs, sorted_suffix
-        )
-    };
-    if is_self_closing {
-        // Self-closing: outer block, optional inner directive block, close both.
-        let opener = format!(
-            " {{{}{}{}{}}}",
-            directive_prefix,
-            inner_open,
-            create(attrs_self),
-            inner_close
-        );
-        str.overwrite(el.start, el.end, &opener);
-    } else {
-        let opener = format!(
-            " {{{}{}{}",
-            directive_prefix,
-            inner_open,
-            create(attrs_open)
-        );
-        str.overwrite(el.start, opening_tag_end, &opener);
-
-        // svelte:element is an element node → children at depth+1.
-        process_fragment_inplace(&el.fragment, source, options, str, counter, depth + 1);
-
-        let closing_tag_start = find_closing_tag_start(source, el.end);
-        let close = format!(" }}{}", inner_close);
-        if closing_tag_start < el.end {
-            str.overwrite(closing_tag_start, el.end, &close);
-        } else {
-            str.append_left(el.end, &close);
-        }
-    }
-
-    // Close the named-slot `$$slot_def[...]` wrapper block; restore context.
-    if named_slot.is_some() {
-        str.append_left(el.end, "}");
-    }
-    counter.slot_inst = saved_slot;
-}
-
-/// Handle `<title>` element.
-fn handle_title_element(
-    el: &TitleElement,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    if el.start >= el.end {
-        return;
-    }
-
-    let opening_tag_end = find_opening_tag_end(source, el.start, el.end);
-    let attrs_str = build_attributes_string(&el.attributes, source, counter.slot_inst.is_some());
-
-    let opener = format!(
-        " {{ svelteHTML.createElement(\"title\", {{{}}});",
-        attrs_str
-    );
-    str.overwrite(el.start, opening_tag_end, &opener);
-
-    // title is an element → children at depth+1.
-    process_fragment_inplace(&el.fragment, source, options, str, counter, depth + 1);
-
-    let closing_tag_start = find_closing_tag_start(source, el.end);
-    if closing_tag_start < el.end {
-        str.overwrite(closing_tag_start, el.end, " }");
-    } else {
-        str.append_left(el.end, " }");
-    }
-}
-
 /// Handle `<slot>` element.
 ///
 /// Generates `{ __sveltets_createSlot("name", { attrs }); fallback_children }`.
@@ -2157,213 +1631,6 @@ fn handle_svelte_self(
     process_fragment_inplace(&el.fragment, source, options, str, counter, depth + 1);
     let trailing = if has_lets { "}}" } else { "}" };
     str.overwrite(closing_tag_start, el.end, trailing);
-}
-
-/// Handle Svelte special elements (svelte:body, svelte:window, etc.).
-///
-/// `svelte:boundary` is special: like `InlineComponent` in the upstream
-/// svelte2tsx, any `{#snippet}` blocks that are **direct children** of
-/// `<svelte:boundary>` become **implicit properties** of the element's
-/// `createElement` attributes object instead of standalone `const` declarations.
-/// This mirrors upstream `SnippetBlock.ts::hoistSnippetBlock` which returns
-/// early for `SvelteBoundary` (treating it exactly like `InlineComponent`),
-/// and `Element.ts::addAttribute` which the upstream `handleSnippet` calls to
-/// insert the snippet body as an attr-value transform.
-///
-/// For all other special elements the snippet children remain standalone
-/// declarations (the default behaviour for elements/blocks).
-fn handle_svelte_special_element(
-    el: &SvelteElement,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    if el.start >= el.end {
-        return;
-    }
-
-    let opening_tag_end = find_opening_tag_end(source, el.start, el.end);
-    let mut attrs_str =
-        build_attributes_string(&el.attributes, source, counter.slot_inst.is_some());
-
-    // Add extra whitespace to match JS svelte2tsx position-preserving behavior
-    if !el.attributes.is_empty() && !attrs_str.is_empty() {
-        let extra_spaces = count_tag_to_attr_spaces(&el.name, el.start, source);
-        if extra_spaces >= 1 {
-            let total_spaces = extra_spaces + 1;
-            let mut padded = " ".repeat(total_spaces);
-            padded.push_str(attrs_str.trim_start());
-            attrs_str = padded;
-        }
-    }
-
-    // `svelte:boundary` treats direct {#snippet} children as implicit props on
-    // the `createElement` attrs object — exactly like InlineComponent in the
-    // upstream. Check whether any direct children are snippet blocks.
-    let has_snippet_children = el.name == "svelte:boundary"
-        && el
-            .fragment
-            .nodes
-            .iter()
-            .any(|n| matches!(n, TemplateNode::SnippetBlock(s) if s.start < s.end));
-
-    if has_snippet_children {
-        // Emit the opener with the attrs object left OPEN so we can append the
-        // implicit snippet props into it before closing. Any regular element
-        // attributes (e.g. `onerror`) come first as normal.
-        //
-        // Result shape:
-        //   { svelteHTML.createElement("svelte:boundary", { <regular-attrs>
-        //     <snippet-name>: (params) => { … return __sveltets_2_any(0) },
-        //   });
-        //   <non-snippet children>
-        // }
-        let opener = format!(
-            " {{ svelteHTML.createElement(\"{}\", {{{}",
-            el.name, attrs_str
-        );
-        str.overwrite(el.start, opening_tag_end, &opener);
-
-        // Process each direct child: transform snippet blocks as implicit props
-        // and move them to anchor (just after the opening tag), then process
-        // non-snippet children in-place (they will appear after the `});`).
-        // Mirrors the `use_snippet_props` branch in `handle_component`.
-        let mut anchor = opening_tag_end;
-        let mut last_snippet_end: Option<u32> = None;
-
-        for node in &el.fragment.nodes {
-            if let TemplateNode::SnippetBlock(s) = node {
-                if s.start >= s.end {
-                    continue;
-                }
-                // Transform the snippet as an implicit attr prop of this
-                // element (same form as a component implicit snippet prop):
-                //   name: (params) => { … return __sveltets_2_any(0) },
-                handle_snippet_block_as_component_prop(s, source, options, str, counter, depth + 1);
-                if s.start == anchor {
-                    anchor = s.end;
-                } else {
-                    str.move_range(s.start, s.end, anchor);
-                }
-                last_snippet_end = Some(s.end);
-            } else {
-                // Non-snippet children live AFTER the createElement call;
-                // svelte:boundary is an ancestor element → depth+1.
-                process_node_inplace(node, source, options, str, counter, depth + 1);
-            }
-        }
-
-        // Close the attrs object and the `createElement(...)` call right
-        // after the last relocated snippet prop.
-        let close_create_element = "});";
-        match last_snippet_end {
-            Some(end) => {
-                str.append_left(end, close_create_element);
-            }
-            None => {
-                // No usable snippet found (shouldn't happen given the guard
-                // above, but guard defensively): close immediately.
-                str.prepend_right(opening_tag_end, close_create_element);
-            }
-        }
-
-        // Close the outer `{ … }` block.
-        let closing_tag_start = find_closing_tag_start(source, el.end);
-        if closing_tag_start < el.end {
-            str.overwrite(closing_tag_start, el.end, " }");
-        } else {
-            str.append_left(el.end, "}");
-        }
-    } else {
-        // `bind:` directives on a special element use the same lowering as a
-        // regular element: `bind:this` and one-way bindings (`clientWidth`, …)
-        // need a `const $$_<name><depth> = createElement(...)` so the binding
-        // assignment (`foo = $$_<name><depth>.clientWidth;` / `target =
-        // $$_<name><depth>;`) can reference it; other two-way bindings get the
-        // generic `() => expr = __sveltets_2_any(null)` widener. Mirrors
-        // upstream Element.ts + Binding.ts.
-        let needs_element_var = any_bind_needs_element_var(&el.attributes, source);
-        let element_var = if needs_element_var {
-            Some(format!("$$_{}{}", element_var_base_name(&el.name), depth))
-        } else {
-            None
-        };
-        let bind_suffix = build_bind_directive_suffix(
-            &el.attributes,
-            source,
-            element_var.as_deref(),
-            &el.name,
-            options.is_ts_file,
-        );
-        let element_var_decl = element_var
-            .as_ref()
-            .map(|v| format!("const {} = ", v))
-            .unwrap_or_default();
-        // `use:` / `transition:` / `animate:` directives on a special element
-        // (e.g. `<svelte:body use:tooltip={…}>`) become the same V4-style
-        // action/transition emission as on a regular element: an
-        // `const $$action_N = __sveltets_2_ensureAction(…);` prefix, a
-        // `__sveltets_2_union($$action_N)` second argument to `createElement`,
-        // and transition/animate suffixes. The action's `mapElementTag` uses the
-        // mapped tag name (`svelte:body` → `body`, per official Element.ts).
-        let action_tag = if el.name == "svelte:body" {
-            "body"
-        } else {
-            el.name.as_str()
-        };
-        let (directive_prefix, directive_suffix, action_count) =
-            build_directive_prefix_suffix(&el.attributes, source, action_tag);
-        let actions_arg = if action_count > 0 {
-            let mut args = String::from(", __sveltets_2_union(");
-            for i in 0..action_count {
-                if i > 0 {
-                    args.push(',');
-                }
-                let _ = write!(args, "$$action_{}", i);
-            }
-            args.push(')');
-            args
-        } else {
-            String::new()
-        };
-
-        // Default path: all children (including any snippets) are processed
-        // as standalone declarations inside the block. When `directive_prefix`
-        // is present it opens an extra outer block scope (for the action
-        // declarations), closed by a matching extra `}` after the children.
-        let opener = if directive_prefix.is_empty() {
-            format!(
-                " {{ {}svelteHTML.createElement(\"{}\", {{{}}});{}{}",
-                element_var_decl, el.name, attrs_str, bind_suffix, directive_suffix
-            )
-        } else {
-            format!(
-                " {{{}{{ {}svelteHTML.createElement(\"{}\"{}, {{{}}});{}{}",
-                directive_prefix,
-                element_var_decl,
-                el.name,
-                actions_arg,
-                attrs_str,
-                bind_suffix,
-                directive_suffix
-            )
-        };
-        str.overwrite(el.start, opening_tag_end, &opener);
-
-        // Special svelte elements (svelte:head, svelte:body, etc.) are element
-        // nodes → children at depth+1, consistent with RegularElement treatment.
-        process_fragment_inplace(&el.fragment, source, options, str, counter, depth + 1);
-
-        let extra_close = if directive_prefix.is_empty() { "" } else { "}" };
-        let closing_tag_start = find_closing_tag_start(source, el.end);
-        if closing_tag_start < el.end {
-            str.overwrite(closing_tag_start, el.end, &format!(" }}{}", extra_close));
-        } else {
-            str.append_left(el.end, &format!("}}{}", extra_close));
-        }
-    }
 }
 
 // =============================================================================
