+++ crates/rsvelte_core/src/svelte2tsx/template/nodes/mod.rs
@@ -6,11 +6,14 @@ pub(super) mod comment;
 pub(super) mod const_tag;
 pub(super) mod debug_tag;
 pub(super) mod declaration_tag;
+pub(super) mod dynamic_element;
 pub(super) mod each_block;
+pub(super) mod element;
 pub(super) mod if_else_block;
 pub(super) mod key_block;
 pub(super) mod mustache_tag;
 pub(super) mod raw_mustache_tag;
 pub(super) mod render_tag;
 pub(super) mod snippet_block;
+pub(super) mod special_element;
 pub(super) mod text;
